pyproject.toml
[project]
name = "guardrag"
version = "0.1.0"
description = "RAG-System auf Basis von COLPALI mit Guardrails für sichere Dokumentenverarbeitung"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "pydantic-ai>=0.4.1",
    "colpali-engine>=0.3.11",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "transformers>=4.45.0",
    "Pillow>=10.0.0",
    "pdf2image>=1.17.0",
    "numpy>=1.24.0",
    "qdrant-client>=1.7.0",
    "aiofiles>=23.0.0",
    "httpx>=0.25.0",
    "tqdm>=4.66.0",
    "reportlab>=4.0.0",
    "pathlib2>=2.3.7",
    "PyMuPDF>=1.23.0",
    "scikit-learn>=1.3.0",
    "sentence-transformers>=2.2.2",
    "opencv-python>=4.8.0",
    "python-magic>=0.4.27",
    "typing-extensions>=4.8.0",
    "fastapi-mcp>=0.1.0",
    "hf-xet>=1.1.5",
    "weasyprint>=65.1",
    "nemoguardrails>=0.14.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.6.0"
]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "httpx>=0.25.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "mcp_fileconverter"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.black]
line-length = 88
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
